Move analyzer executables to separate package/files
